@@ -1,8 +1,10 @@
-import 'package:dart_application_1/dart_application_1.dart' as dart_application_1;
+import 'package:dart_application_1/dart_application_1.dart'
+    as dart_application_1;
 
 void main(List<String> arguments) {
   print('Hello world: ${dart_application_1.calculate()}!');
 
   double temp = 38.457;
-  String tempText = temp
+  // ignore: unused_local_variable
+  double tempText = temp;
 }
